Move internal/lazyregexp to daemon/internal/lazyregexp
