+++ daemon/builder/dockerfile/dispatchers_windows.go
@@ -8,7 +8,7 @@ import (
 	"path/filepath"
 	"strings"
 
-	"github.com/docker/docker/internal/lazyregexp"
+	"github.com/docker/docker/daemon/internal/lazyregexp"
 	"github.com/moby/buildkit/frontend/dockerfile/instructions"
 	"github.com/moby/moby/api/types/container"
 )
