+++ daemon/events/testutils/testutils.go
@@ -5,7 +5,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/docker/docker/internal/lazyregexp"
+	"github.com/docker/docker/daemon/internal/lazyregexp"
 	"github.com/moby/moby/api/types/events"
 	timetypes "github.com/moby/moby/api/types/time"
 )
