+++ daemon/internal/lazyregexp/lazyregexp.go
