+++ daemon/internal/lazyregexp/lazyregexp_test.go
