+++ daemon/logger/awslogs/cloudwatchlogs.go
@@ -23,10 +23,10 @@ import (
 	smithymiddleware "github.com/aws/smithy-go/middleware"
 	smithyhttp "github.com/aws/smithy-go/transport/http"
 	"github.com/containerd/log"
+	"github.com/docker/docker/daemon/internal/lazyregexp"
 	"github.com/docker/docker/daemon/logger"
 	"github.com/docker/docker/daemon/logger/loggerutils"
 	"github.com/docker/docker/dockerversion"
-	"github.com/docker/docker/internal/lazyregexp"
 	"github.com/pkg/errors"
 )
