+++ daemon/names/names.go
@@ -1,6 +1,6 @@
 package names
 
-import "github.com/docker/docker/internal/lazyregexp"
+import "github.com/docker/docker/daemon/internal/lazyregexp"
 
 // RestrictedNameChars collects the characters allowed to represent a name, normally used to validate container and volume names.
 const RestrictedNameChars = `[a-zA-Z0-9][a-zA-Z0-9_.-]`
