+++ daemon/pkg/opts/opts.go
@@ -7,7 +7,7 @@ import (
 	"path"
 	"strings"
 
-	"github.com/docker/docker/internal/lazyregexp"
+	"github.com/docker/docker/daemon/internal/lazyregexp"
 	"github.com/docker/go-units"
 )
